feat(examples): Added Example App
This change adds an example app to the repo which is served using webpack dev server and renders two
tile layers (OSM and XYZ) which can be switched between. The change includes the addition of an OSM
tile layer class, and an example folder under source containing the example app.