fix(tests): Fixed Pinch Tests
With the upgrade to open layers 5.3.0 some new properties were added to the pinch interactions (more
generally the PointerInteractions). Currently the tests for the pinch interactions was doing a
direct comparison between their call and the object that they get called with. I changed this to use
"expect.objectContaining" so that other properties are ignored.